refactor: move deal parsing code for reuse
